:sparkles: Introduce a serverless token auth proxy, fixes #20
We're using a serverless approach to GitHub GraphQL API authentication
thanks to a Cloud Function. It acts as a proxy to the API handling the
Personal Access Token on the server side. The serverless function should
be configured with `GITHUB_GRAPHQL_API_URL` and `GITHUB_PAT` and is
deployed using Terraform.
The serverless function is currently deployed to:
https://us-east1-gitpop3.cloudfunctions.net/gitpop3-graphql